+++ tests-unit/string_literal_connect.in.c
@@ -1,0 +1,5 @@
+#define STRING_IN_MACRO "nana"
+main () {
+	// String without commas connect to one another
+	echo("" "banana" "na" STRING_IN_MACRO);
+}