+++ tests-unit/string_literal_connect.out.c
@@ -1,0 +1,11 @@
+var __addr;
+var __rval;
+var __sp;
+
+main
+{
+  __sp = 512;
+  label __main_loop:
+  echo('bananananana');
+  goto __main_loop;
+}